change: sniper movement effects crit and damage
fix: set ReloadTimeMs type to float to fix loss of precision bugs
change: remove crosshair spread on sniper shoot

+++ src/Items/Offensive/FirearmStats.cs
@@ -8,7 +8,7 @@ public partial class FirearmStats : BaseOffensiveStats
 	public int ProjectileRadius = 24;
 
 	[Export]
-	public int ReloadTimeMs = 1500;
+	public float ReloadTimeMs = 1500;
 
 	[Export]
 	public float BloomCoefficientDeg = 0.03f;
